+++ vfrecovery/downloaders/__init__.py
@@ -1,2 +1,3 @@
 from armor3d import Armor3d
 from glorys import Glorys
+from core import get_velocity_field
